+++ src/Telegram.Bot/Types/Enums/MessageType.cs
@@ -71,7 +71,12 @@ namespace Telegram.Bot.Types.Enums
         /// <summary>
         /// The <see cref="Message"/> contains a <see cref="Game"/>
         /// </summary>
-        GameMessage
+        GameMessage,
+
+        /// <summary>
+        /// The <see cref="Message"/> contains a <see cref="VideoNote"/>
+        /// </summary>
+        VideoNoteMessage,
     }
 
     internal static class MessageTypeExtension
@@ -102,6 +107,8 @@ namespace Telegram.Bot.Types.Enums
                     return new KeyValuePair<string, string>("sendVenue", "latitude");
                 case MessageType.GameMessage:
                     return new KeyValuePair<string, string>("sendGame", "game_short_name");
+                case MessageType.VideoNoteMessage:
+                    return new KeyValuePair<string, string>("sendVideoNote", "video_note");
 
                 default:
                     throw new NotImplementedException();